Рефакторинг сервера и агента + немного тестов
