+++ .github/workflows/mertricstest.yml
@@ -83,7 +83,7 @@ jobs:
           metricstest -test.v -test.run=^TestIteration1$ \
             -binary-path=cmd/server/server
 
-      - name: Build agent binary
+      - name: Build client binary
         run: |
           cd cmd/agent
           go build -buildvcs=false  -o agent
